Get Function Release
This Get Function is in release form

It should work elegantly with common scenarios. I haven't tested the extremely odd scenarios but it should work generally.

I've recieved some wierd behavior and made some work arounds.
Particularly at line 525 and 567  in end and finish function.

For line 525, the  inhouse_players.append(inhouse_t2.pop()) causes a dupe.

For line 567,
channel = self.bot.get_channel(inhouse_channels.get("Empty"))
This is required because deleting the list causes inhouse_players to break with the get function. I'm not sure why.
What happens is that if you try to use the get function again, it will return nothing for the player list. This workaround points to an always empty server (as its locked to all) . This will empty the list instead of deleting it.